+++ packages/doenetml-worker-rust/lib-doenetml-core/src/core/dast/flat_dast.rs
@@ -1,2 +1,5 @@
-pub mod parent_iterator;
-pub mod untagged_flat_dast;
+mod parent_iterator;
+mod untagged_flat_dast;
+mod untagged_flat_dast_merge;
+
+pub use untagged_flat_dast::*;
