+++ packages/doenetml-worker-rust/lib-doenetml-core/src/core/dast/flat_dast/untagged_flat_dast_merge.rs
@@ -1,0 +1,231 @@
+use crate::dast::{
+    DastElement, DastElementContent, DastError, DastFunctionMacro, DastMacro, DastRoot,
+    DastTextMacroContent,
+};
+
+use super::{
+    FlatAttribute, FlatElement, FlatError, FlatFunctionMacro, FlatMacro, FlatNode, FlatRoot, Index,
+    ParentIterator, UntaggedContent,
+};
+
+impl FlatRoot {
+    /// Create a new `FlatRoot` from a `DastRoot`.
+    pub fn from_dast(dast: &DastRoot) -> Self {
+        let mut flat = Self::new();
+        flat.merge_dast_root(dast);
+        flat
+    }
+
+    /// Iterate over the parent elements of a node.
+    /// If for some reason the node has a non-element parent, the iterator will panic.
+    pub fn parent_iter(&self, start_idx: Index) -> ParentIterator {
+        let start = &self.nodes[start_idx];
+        ParentIterator::new(start, self)
+    }
+
+    /// Merge the content of a `DastRoot` into `FlatRoot`.
+    /// This function recursively adds all children etc. of the `DastRoot`.
+    pub fn merge_dast_root(&mut self, dast: &DastRoot) {
+        for child in dast.children.iter() {
+            self.merge_content(child, None);
+        }
+    }
+
+    /// Update the the node at `idx` to be `node`. Any children of `node` are added to the `FlatRoot` as well,
+    /// but children of the old node are not removed. This function does no consistency checking, so it is
+    /// the _caller should be extra careful_.
+    pub fn update_content(
+        &mut self,
+        node: &DastElementContent,
+        idx: Index,
+        parent: Option<Index>,
+    ) -> UntaggedContent {
+        match node {
+            DastElementContent::Element(elm) => {
+                self.set_element(elm, idx, parent);
+                let ret = UntaggedContent::Ref(idx);
+                // Newly set elements have empty children and attributes. It is our
+                // job to fill them in.
+                let children = elm
+                    .children
+                    .iter()
+                    .map(|child| self.merge_content(child, Some(idx)))
+                    .collect();
+                let attributes: Vec<FlatAttribute> = elm
+                    .attributes
+                    .values()
+                    .map(|attr| {
+                        let children = attr
+                            .children
+                            .iter()
+                            .map(|child| match child {
+                                DastTextMacroContent::Text(txt) => {
+                                    DastElementContent::Text(txt.clone())
+                                }
+                                DastTextMacroContent::Macro(macro_) => {
+                                    DastElementContent::Macro(macro_.clone())
+                                }
+                                DastTextMacroContent::FunctionMacro(function_macro) => {
+                                    DastElementContent::FunctionMacro(function_macro.clone())
+                                }
+                            })
+                            .map(|child| self.merge_content(&child, Some(idx)))
+                            .collect();
+                        FlatAttribute {
+                            name: attr.name.clone(),
+                            parent: Some(idx),
+                            children,
+                            position: attr.position.clone(),
+                        }
+                    })
+                    .collect();
+                self.set_children(idx, children);
+                self.set_attributes(idx, attributes);
+                ret
+            }
+            DastElementContent::Error(err) => {
+                let ret = UntaggedContent::Ref(self.set_error(err, idx, parent));
+                if parent.is_none() {
+                    self.children.push(ret.clone());
+                }
+                ret
+            }
+            DastElementContent::Text(txt) => UntaggedContent::Text(txt.value.clone()),
+            DastElementContent::Macro(macro_) => {
+                UntaggedContent::Ref(self.set_macro(macro_, idx, parent))
+            }
+            DastElementContent::FunctionMacro(function_macro) => {
+                self.set_function_macro(function_macro, idx, parent);
+                let ret = UntaggedContent::Ref(idx);
+                if let Some(input) = function_macro.input.as_ref() {
+                    let input: Vec<Vec<UntaggedContent>> = input
+                        .iter()
+                        .map(|arg| {
+                            arg.iter()
+                                .map(|arg| self.merge_content(arg, Some(idx)))
+                                .collect()
+                        })
+                        .collect();
+
+                    self.set_function_macro_input(idx, input);
+                }
+                ret
+            }
+        }
+    }
+
+    /// Merge DAST content into `FlatRoot`.
+    pub fn merge_content(
+        &mut self,
+        node: &DastElementContent,
+        parent: Option<Index>,
+    ) -> UntaggedContent {
+        let ret = match node {
+            DastElementContent::Text(_) => {
+                // A text node doesn't get pushed to `nodes`, so the value of `idx` doesn't matter.
+                // We pass in `Index::MAX` so that an error will be thrown if this value is used.
+                self.update_content(node, Index::MAX, parent)
+            }
+            _ => {
+                // We push an error onto the nodes stack and then we do an in-place update of
+                // that error to the correct node. This avoids code duplication.
+                let idx = self.nodes.len();
+                self.nodes.push(FlatNode::Error(FlatError::with_message(
+                    "TEMPORARY NODE CREATED DURING `merge_content()`".to_string(),
+                    idx,
+                )));
+                self.update_content(node, idx, parent)
+            }
+        };
+
+        if parent.is_none() {
+            self.children.push(ret.clone());
+        }
+        ret
+    }
+
+    /// Set the children of an element node.
+    pub fn set_children(&mut self, idx: usize, children: Vec<UntaggedContent>) {
+        match &mut self.nodes[idx] {
+            FlatNode::Element(elm) => {
+                elm.children = children;
+            }
+            _ => panic!("set_children called on non-element node"),
+        }
+    }
+
+    /// Set the attributes of an element node.
+    fn set_attributes(&mut self, idx: usize, attributes: Vec<FlatAttribute>) {
+        match &mut self.nodes[idx] {
+            FlatNode::Element(elm) => {
+                elm.attributes = attributes;
+            }
+            _ => panic!("set_attributes called on non-element node"),
+        }
+    }
+
+    /// Set the input of a function macro node.
+    fn set_function_macro_input(&mut self, idx: usize, args: Vec<Vec<UntaggedContent>>) {
+        match &mut self.nodes[idx] {
+            FlatNode::FunctionMacro(func_macro) => {
+                func_macro.input = Some(args);
+            }
+            _ => panic!("set_function_macro_args called on non-function-macro node"),
+        }
+    }
+
+    /// Set the node at position `idx` to be the specified element.
+    /// The element will be initialized with empty children and attributes.
+    fn set_element(&mut self, node: &DastElement, idx: Index, parent: Option<Index>) -> usize {
+        self.nodes[idx] = FlatNode::Element(FlatElement {
+            name: node.name.clone(),
+            children: Vec::new(),
+            attributes: Vec::new(),
+            position: node.position.clone(),
+            parent,
+            idx,
+            // It is impossible to directly set `extending` in DAST; it is computed later.
+            extending: None,
+        });
+        idx
+    }
+
+    /// Set the node at position `idx` to be the specified error.
+    fn set_error(&mut self, node: &DastError, idx: Index, parent: Option<Index>) -> usize {
+        self.nodes[idx] = FlatNode::Error(FlatError {
+            message: node.message.clone(),
+            position: node.position.clone(),
+            parent,
+            idx,
+        });
+        idx
+    }
+
+    /// Set the node at position `idx` to be the specified macro.
+    fn set_macro(&mut self, node: &DastMacro, idx: Index, parent: Option<Index>) -> usize {
+        self.nodes[idx] = FlatNode::Macro(FlatMacro {
+            path: node.path.clone(),
+            position: node.position.clone(),
+            parent,
+            idx,
+        });
+        idx
+    }
+
+    /// Set the node at position `idx` to be the specified function macro.
+    fn set_function_macro(
+        &mut self,
+        node: &DastFunctionMacro,
+        idx: Index,
+        parent: Option<Index>,
+    ) -> usize {
+        self.nodes[idx] = FlatNode::FunctionMacro(FlatFunctionMacro {
+            path: node.path.clone(),
+            input: None,
+            position: node.position.clone(),
+            parent,
+            idx,
+        });
+        idx
+    }
+}
